+++ src/client.ts
@@ -1,4 +1,5 @@
 import type { Stream } from "./stream.js";
+import type { IntMode } from "./value.js";
 
 export type ProtocolVersion = 1 | 2;
 export const protocolVersions: Map<string, ProtocolVersion> = new Map([
@@ -9,7 +10,9 @@ export const protocolVersions: Map<string, ProtocolVersion> = new Map([
 /** A client for the Hrana protocol (a "database connection pool"). */
 export abstract class Client {
     /** @private */
-    constructor() {}
+    constructor() {
+        this.intMode = "number";
+    }
 
     /** Get the protocol version negotiated with the server. */
     abstract getVersion(): Promise<ProtocolVersion>;
@@ -22,4 +25,11 @@ export abstract class Client {
 
     /** True if the client is closed. */
     abstract get closed(): boolean;
+
+    /** Representation of integers returned from the database. See {@link IntMode}.
+     *
+     * This value is inherited by {@link Stream} objects created with {@link openStream}, but you can
+     * override the integer mode for every stream by setting {@link Stream.intMode} on the stream.
+     */
+    intMode: IntMode;
 }